5.1.0

+++ Pod/Proximiio.framework/Headers/ProximiioGeoJSON.h
@@ -30,7 +30,7 @@
 
 # pragma mark - Feature
 + (NSArray *)features;
-+ (int) lastDbSync;
++ (NSInteger) lastDbSync;
 + (void) setLastDbSync:(int) time;
 + (ProximiioGeoJSON *)featureInitWithJson:(NSDictionary *)json;
 + (ProximiioGeoJSON *)featureLineWithCoordinates:(NSArray *)coordinates properties:(NSDictionary *)properties;
